+++ src/contexts/auth/useAuth.ts
@@ -1,5 +1,5 @@
 import { useContext } from 'react'
 
-import { AuthContext } from './authContext'
+import { AuthContext } from './AuthContext'
 
 export const useAuth = () => useContext(AuthContext)
